change repository

+++ src/main/java/com/ms/data/master/campaign/model/dto/campaign/CampaignScanSummaryProjection.java
@@ -7,7 +7,7 @@ public interface CampaignScanSummaryProjection {
 
     String getProductName();
 
-    Long getTotalScan();
+    Integer getTotalScan();
 
     Integer getScanPerClaim();
